Remove Async Postgres Operator
The Async Postgres Operator is not reliable currently as it does not work with the case where the SQL query is correct but the column names and table names are "non-existent".

So we won't be able to reliably run this operator.

Submitting queries directly from Triggerer will cause issues when multiple Triggerers are running and one of them lose DB connection, at the point duplicate triggers would be running. So operations like Submitting computationally heavy queries or update/edit/delete based queries can cause havoc. And there is no way to guarantee that the user has submitted an idempotent query.
Secondly, a trigger can fail and Triggerer will restart it which can cause the same issue.