Handle  more  blocking code and get test passing
There was some blocking code in super().__init__ that wasn't being awaited.  Namely a call to get_connection.  It was tricky to figure out how to get around  this, but  this  is  what I came up with.

Additionally,  the test previously assumed  that a  connection existed  for google so I mocked that out.

Finally I figured out how to mock that session context manager properly so the test will pass (which was the original objective.)