Automate releasing artifacts to PyPI
This PR/commit automates the process of publishing the artifacts to PyPI. The publish and build step only runs when  a new tag is created. This PR also includes some verifications like "version" in setup.cfg matches the Git Tag and then Changelog contains entries for this version (for a non-dev).