Automate inventory list for all providers
This commit autogenerates the list using the list of providers in "all" extras in setup.cfg. This way we don't need to manually add providers when we add operators for new Providers. This also fixes the issue with missing Livy inventory.

This helps in the docs as shown below by allow users to click on the link for the docs of Base/inherited class.

<img width="1043" alt="image" src="https://user-images.githubusercontent.com/8811558/163276645-dda3d0af-c222-44a2-984e-518d164b01f8.png">

But because Livy doesn't have an inventory, it doesn't show right now:
<img width="771" alt="image" src="https://user-images.githubusercontent.com/8811558/163276832-e5b12282-119b-4318-a8f2-fefb883dd513.png">