Fail the shell script if any of the commands fail
Currently, the shell script is silently ignoring failure of
commands. As a result, the Airflow task calling this shell script
is marked as 'success' which should not happen because the DAG
proceeds to execute subsequent tasks resulting in late failures.
With this change, we want to report error in any command upfront
so that the task is marked 'failed' and the DAG does not continue
executing downstream tasks.