Pass needed environment variables to shell script
The shell script to create and delete EKS cluster along with other
resources needs a set of environment variables which were not
being passed from the respective Airflow operators. These are being
passed now.
The change also orders the variables alphabetically.

Additionally, the 'create_emr_virtual_cluster' task generates
the cluster ID which needs to be passed to subserquent tasks. For
this purpose, we push the ID to xcom and the subsequent tasks pull
it from the XCOM database.