Fix ingress port authorizations for Livy & Hive example DAGs (#342)
* Fix ingress port authorisations for Livy & Hive example DAGs

The previous implementation was checking only for the presence
of the machine's IP address for existence of ALLOW permission
in the security group. As a result, when both Livy and Hive example
DAGs run concurrently, one of them adds the machine's IP along with
its ports to be authorised, but the second DAG sees that the IP
is already added and then skips authorising its port. To elaborate
with an example, let's suppose the Hive example DAG runs first. It
will authorise and open port '22' and '10000' for the machine's IP.
When the Livy operator runs, it will see that rules against the
machine's IP already exist in the security group and as a result
it will skip authorising its required port '8998' in the security
group.

To fix this, the commit removes the check on IP in the security
group and instead tries authorising the ports directly. In case,
the rule already exists boto will throw a duplicate permission error
which we catch. Other errors are reported back to ensure that the
task is marked failed. Previously, since, the port was getting
skipped to be open, we were getting the error mentioned in issue #298