Loosen Airflow version pin to >2.2.0 (#345)
Apparently, there was some db upgrade issue on my setup which
I thought is due to version mismatch in base Docker image and
the airflow dependency version. That however is not the case
as it works fine with >=2.2.0 and we do not need to pin it to >=2.3.0.
Pinning it to >=2.3.0 would mean users on <2.3 won't be able to use
the providers packages. Hence, reverting it to >=2.2.0