Reconstruct hook using connection ID
I observed that the hook that we attach in `execute` method
https://github.com/astronomer/astronomer-providers/blob/main/astronomer/providers/google/cloud/operators/bigquery.py#L92
is not available in `execute_complete` https://github.com/astronomer/astronomer-providers/blob/main/astronomer/providers/google/cloud/operators/bigquery.py#L129
I tried to attach this hook in `execute_complete`; it does get
attached, however, it no longer remains attached to the operator
outside its scope. Previously, the hook attached in `execute` was
available throughout (even in the extractor), but this no longer
seems to be the case. Something has changed internally which is
wiping our the hook from the operator that we manually attach to it.
To make the connection available in the extractor, I am reconstructing
it using the hook utility and this seems to fix the issue.

I am creating this commit to fix the current issue and
what has change internally needs to be looked into for understanding
purpose.