Separate out Dockerfiles for Gen1 and Astro Cloud image builds
It is observed that even if we intend to build image for Astro Cloud
with target in the integration-test script, it still tries to create
the earlier image of Gen1. For this reason and also to keep separate
contexts, the commit separates out the corresponding Dockerfile.
This will help us in saving build time.