List installed pip packages used for running master DAG in Integration Tests deployment (#486)
* Add BashOperator task in master DAG to list installed python packages

It is good to see what all python packages are installed with the
build so as to understand what versions are being used by the DAGs
run as part of the master DAG in Astro Cloud.

* Separate out Dockerfiles for Gen1 and Astro Cloud image builds

It is observed that even if we intend to build image for Astro Cloud
with target in the integration-test script, it still tries to create
the earlier image of Gen1. For this reason and also to keep separate
contexts, the commit separates out the corresponding Dockerfile.
This will help us in saving build time.