Fix example_async_bigquery_queries dependency
It is observed that the `delete_dataset` task is called even before
the dependent tasks finish processing which depend on the dataset
to be queried upon. The commit fixes this issue by adding appropriate
dependecny in the DAG chain.

Also, since the `delete_dataset` task is called with the `trigger_rule`
`all_done`, we're adding an end task to report the DAG status correctly,
i.e. fail the DAG if any of the tasks fail.