Add new Great Expectations Provider
New operator is added with limited testing, as it is a first draft
that may not be using certain elements of the GE project in an
efficient or correct manner. It is prudent to have the operator
cleaned up first before high-coverage testing is added to functions
that may change significantly or be removed. Additionally, no
example DAG is yet prepared as parameters to the operator may
change significantly.