Upgrade base Airflow image to use latest release (#509)
* Upgrade base Airflow image to latest release

astro-runtime:5.0.6 and ap-airflow:2.3.3-1 were released on
11th July 2022. The commit updates our Dockerfiles to reference these
as the base images.

The commit also makes changes to have a common variable for make targets to
use the image name. With this, going forward, we will need to update the images
in only 2 places now, 1 for the make targets and 1 for the CircleCI deployment image.