Fix async HTTP sensor failing with http: 4.0.0
The http: 4.0.0 provider release removes the assignment of hook
variable to the instance. As a result, the derived class
HttpSensorAsync extending the base class HttpSensor fails as it
relies on the base class' instance variable hook. The fix involves
reinstantiating the hook so that it is available for usage in the
subsequent methods.