Add 15 mins sleep between deploying image & triggering master DAG
It is observed on Astro Cloud that when a new image is deployed it
takes certain time to update the image on all Airflow component pods.
Currently, there is no programatic way to know whether the image
has been updated across all components. Hence, once the image is
deployed as part of the RC test make target, we are introducing
a sleep of 15 mins to allow the image to be updated across all
Airflow component pods before triggering the master DAG.