Add temporary logs to debug intermittent failures of the Dataproc DAG (#583)
The Dataproc DAG has been failing intermittently due to timeout
error. The Timeout error is a custom error returned by the trigger
when the end_time value passed to the trigger is greater than the
default timeout value of 3600 seconds set when the cluster creation
is triggered. We want to check why is the end_time condition check
failing for the DAG by checking its set value when the DAG fails.