Modify test RC dependency automation to create a PR
Currently, we only modify the setup.cfg with the RC providers,
build the image, deploy it to Astro Cloud and then trigger the
master DAG run. However, we do not check if the RC causes any
static check failures like mypy or if the tests run fine as we
do not run the checks which are only triggered once we create a PR.
We come to know of such failures only upon new subsequent PRs.
So to detect such failures early, change the test-rc-deps make target
to create a PR by pushing the setup.cfg to a new branch.

Also, we observed in the past that it takes sometimes ~20 mins for
the new image changes to be applied on Astro Cloud and hence we
are increasing the wait time to 30 mins before triggering the master
DAG run to test the RCs against our DAGs.