Common trigger-operator combo for ping-pong
Instead of running a loop in the trigger, this essentially "uses" the
triggerer process as the loop. The idea is:

1. Operator gets deferred after submitting the query.
2. The trigger waits and checks the result *once* and reports.
3. The operator is resumed to check the status
    3a. If the job is reported as still running, defer to run the
        trigger *again*. Go back to 2.
    3b. If the job is done (success or error), return or raise.

This kind of achieves the ping-pong effect and emit the relevant logs in
the operator's worker process, instead of in the triggerer process. The
repeated triggering would probably have some overhead, but should be OK
since everything is async.

Experiementing first on BigQueryInsertJobOperatorAsync; will apply to
others afterwards.