Share cache for pip and pyenv installations across branches
Since we calculate the checksum of the setup.cfg and build unique
keys with this checksum for storing cache, I believe we should share
the cache across branches as long as there is no setup.cfg changes.
This helps us reduce CI storage that is created for each branch
and additionally speeds up CI run for every new branch if there is
no change in the setup.cfg file.