Support retriveing extra fields which do not have extra prefix set
With the 5.0.0 release of the Microsoft Azure PR, the PR https://github.com/apache/airflow/pull/27758
introduced a breaking change where while creating new connections,
the ``extra__`` prefix is no longer set for extra fields in the conneciton.
This issue was not identified with testing the 5.0.0 RC because, it
only happens for new connections that are created. The existing connections
still contain the extra fiels with the ``extra__`` prefix. Hence, the
existing code which looks for the connection field with the prefix
``extra__azure_data_factory__subscriptionId`` works on the older deployment
with the new provider release as the connection was created before the release,
but it fails on new deployments when a fresh connection is created.

To fix this, we're removing the prefix while retrieving the connection
field and at the same time we're supporting previously created connections
by using the same ``get_field`` method from Airflow OSS introduced
in the same PR above to allow backward compatibility.