Set TriggerDagRunOperator's trigger_rule to all_done in master_dag (#950)
The master_dag has implemented chaining of DAGs by grouping them
based on providers. However, if one of the DAG fails in the chain,
subsequent DAGs are not run and hence we do not test them. As such,
the DAGs do not need to depend on each other and hence we can set
the trigger_rule in the TriggerDagRunOperator to all_done so that
we test all the DAGs irrespective of whether the previous DAG in
the chain succeeded.