Support host key verification for SFTPSensorAsync
Since, the AsyncSSH client does not support host keys other than
the ones in known_hosts file, we create and use a paramiko client
to verify the provided host key. If no host key is given, we issue
a warning in the log that it is succeptible to man in the middle
attacks.