Use runtime image 8.0.0-alpha1 so that nightly job succeeds
We recently deployed 8.0.0-alpha1 to our provider-integartion-tests
deployment. The current nightly job tries to deploy 7.4.2 image, but
it fails as Astro does not allow to downgrade the runtime image for
a deployment. So, until the stable runtime image is out, let's deploy
the alpha-1 image so that we ensure there are no failures as there
are multiple provider releases in short span of time.