Schedule daily run of AWS nuke DAG temporarily (#1040)
With the change in PR #1019, we set the schedule of nuke DAG to
none as it runs from the master DAG now. However, there are some
abnomalities in the KubernetesExecutor deployment and also the
nuke DAG does not run there. As a result resources are not cleaned
up post the master DAG run in KubernetesExecutor (which runs after
the master DAG in CeleryExecutor). As a result, on next day, DAGs
in CeleryExecutor deployment also fail due to resouce already
exists error. Hence, for the time being until KubernetesExecutor
deployment is all well, we would like to renable the Nuke DAG to run
daily at 20:30 Hrs.