Combine aws configure with aws nuke in single operator (#1046)
It is observed on KubernetesExecutor deployment that the AWS Nuke DAG fails
with permission issues that it cannot assume an AWS role to perform operations
in our AWS account. The preliminary guess is that since we run authentication
and nuke commands in separate operators, they get launched in separate pods
and hence the authentication/authorization is missing the second pod which triggers
the nuke. This PR combines the two operators in a single operator so that the Nuke
DAG runs fine on our KubernetesExecutor deployment