Mark SageMaker DAG failed if any of the tasks fail (#1054)
Currently, as seen in the issue the DAG is marked success based
on the status of only the last task even if any of the intermediate tasks
fail. As a result, the master DAG does show the status of the DAG as
successful even though its intermediate tasks fail. This does not reflect
the correct top-level view of the statuses of the DAGs triggered from the master DAG.
The commit adds a PythonOperator task at the end of the DAG to check the
statuses of all of the DAG's tasks and marks the DAG as failed if any of its tasks
fail.

closes: #1053