Use `self.job_id` to describe Batch job
With the change in PR #https://github.com/astronomer/astronomer-providers/pull/1087,
we called `submit_job` and assigned its output to a local variable
`job_id` and then used this local variable `job_id` to describe on the
job subsequently by passing it as parameter to the describe method.
But, the submit_job method implicitly assigns the `job_id` as an
instance attribute and returns None. Hence, `self.job_id` actually
contains the `job_id` and the local variable `job_id` which was capturing
the output of `submit_job` was assigned `None` value and hence the
subsequent method to describe job was failing as it was calling
the describe on a None (job_id) value.

The commit fixes this issue by referencing `self.job_id` instead
of using the `None` valued local variable `job_id`

closes: #1097.