Enhance SFTP Async sensor addressing user feedback (#1072)
The PR address the following feature requests mentioned in the issue as per user feedback:

If the ssh host key and the SSH command line tool outputs are in the format of ecdsa-sha2-nistp256, then ideally the SFTP hook should recognise this format without the need for it to be edited (by prepending ssh-) just to get it working with Astronomer’s SFTP library
If no_host_key_check: False on the connection, but doesn’t include a host key, it lets the connection through just fine and effectively doesn’t validate the server. This seems wrong. In fact, the whole need to enable host key checking by default seems wrong. Host key check should be enabled by default (so by default users are then forced to use secure/validated connections), and if it is needed to test locally without ssh host key checking, let the user turn on no_host_key_check: True in the connection details to override the default. It seems a safer approach.
Users are still having to write out to the docker containers ~/.ssh/config file as part of their work because one of the servers they connect to is using a much older key exchange algorithm that the ssh libraries don’t check for by default, so they have to tell it which one to use for a specific site. Given they still need to write out this file, it would be convenient for them just to update the ~/.ssh/known_hosts file too, now they know how to do it as part of the Dockerfile. It would make sense to them for the SFTP provider to default back to using the known_hosts file if a host key is not specifically provided on the connection. Developers have the choice then of which approach to take.
Accept string values(fetched from XCOM results of the previous task) to the newer_than field that need to be converted to datetime values in the implementation as currently the newer_than field only accepts datetime fields.
Use readdir instead of listdir to avoid multiple roundtrip for each file to get the modified timestamp of the file.
Based on code review suggestion by @ashb that we could pass byte string known_hosts to the asyncssh client connection for the given connection extras host_key, we do not need an additional paramiko based host key validation as the asynchssh client usage can validate host keys by itself. This also removes the need for the presence of known_hosts files in case of host_key provided via connection extras. Thus, we also clean up the paramiko based validation added in the previous PR #963

closes: #1021

---------

Co-authored-by: Ash Berlin-Taylor <ash_github@firemirror.com>