Modify constraints job to run only on release and main branches (#1129)
Currently, we run the generate-constraints job in CircleCI for a matrix of
Python and Airflow versions (total 9 jobs at the moment) for every PR that
we create and any additional commits we push to the PR also trigger that
job. This is unnecessary.
The PR modifies the generate-constraints job to be run only on the main
and release (branches starting with the release prefix) branches. It also
updates the RELEASING.rst to suggest that release branches need to have
that practice although it's unconventional not to do so.