Fix additional mypy failures newly reported in main
PR #1143 showed in its checks that all mypy check is successful,
but after merging to main, main reported additional couple of failures.
The commit tried to resolve those mypy failures too by ignoring them
as we do mean to override the methods as they are being implemented
in the respective providers.