Terminate running EMR containers virtual clusters
It has been observed that some times the virtual clusters launched
by EMR containers are not cleaned up. It is not possible to nuke
them using AWS nuke as this resource type is not exposed, additionally
it has not been exposed to delete them in the AWS console too. The
only way to delete them is using AWS CLI and hence we're adding
a task in the nuke DAG to delete those using AWS CLI.