Update the dev AWS Dockerfile to support all AWS example DAGs (#1152)
We use the dev/Dockerfile.aws to build images locally for testing
AWS example DAGs. Images built using this file currently do not
support running example DAGs like the nuke dag, emr_eks_pi_job_dag
due to missing dependencies. The commit adds the missing dependencies.
Additionally, since most of the developers are on ARM64 architectures,
it makes sense to update the AWS CLI installation to refer to aarch64
binary instead of the amd64; so we're also changing that in the commit.