Remove AWS nuke DAG from master DAG and accept DAG schedule param for it
It is observed that when 2 deployments are run in parallel, their
master DAGs trigger the nuke task which cleanup the resources and
if one master DAG finished running before the other one across the
deployments, the other deployment's master DAG tasks will fail as
the resources will no longer be available. To mitigate this, we
remove the nuke DAG trigger from the master DAG and instead plan
to schedule the master DAG only from one of the provider integration
tests deployment (e.g. CeleryExecutor) by using the env var
"AWS_NUKE_DAG_SCHEDULE". We also want to be alerted in our Slack
channel if one of the nuke tasks fail and hence also add a report
task for the DAG.