Fix run of deleting stale VPCs on Astro Cloud
The Astro Cloud Dockerfile for integrations tests copies files
having a prefix of `example_` only in the image getting built.
The script ran fine locally because, in the local dev image, using
docker-compose we mount the entire dev/dags directory.
The PR fixes this issue for Astro Cloud by adding the required
prefix and then calling the renamed file from the DAG.