Increase wait timeout for EKS cluster deletion to 60 minutes
Previously, we were just triggering the delete of the EKS cluster
without waiting for it to finish, as a result the task did not
reflect the accurate status of whether the EKS cluster got deleted.
However, in PR #1157, we introduce the `--wait`` flag on the delete
command so that the Airflow task reflects the correct status.
However, it has been observed that the deletion takes longer than
the default 25m0s timeout set by the command. Hence, we're increasing
the timeout to 60m0s. I think this is safe number given that we also
run the Nuke DAG in some time after the master DAG run which should
clean up leftover resources.
Additionally, we want to continue the deletion of the cluster in case
of failures as there are no dependencies on the cluster and hence
we're also introducing the `--force` flag while deletion.
I believe with the addition of these 2 flags, the task should not
fail again.