Extend upstream Amazon provider SageMaker operators; set deferrable to True (#1169)
The PR aims at extending the upstream SageMaker operators of the 
Amazon providers of Apache Airflow for the asynchronous SageMaker
operators in ``astronomer-providers``, setting the `deferrable` param
to `True` while initialising the corresponding parent constructor and 
cleaning up the code for the existing implementation.

With this, the aim is to push support and fixes to the upstream operators 
as much as possible, and also we do not have to keep track of the 
upstream changes and backport them to the operators in this module.

Following is the high-level summary of the changes made in the PR:
1. Modify the `astronomer/providers/amazon/aws/operators/sagemaker.py`
module to incorporate an extension of corresponding upstream operators in the 
asynchronous operators and set `defferable=True` while initialising the parent
class constructor. Since the parent class supports deferrable capabilities well
now, we do not need any additional code for the operators here; so we clean 
that up here
2. Remove the following modules as they are no longer needed. They are 
maintained in the upstream provider.
i.  `astronomer/providers/amazon/aws/hooks/sagemaker.py`
ii. `astronomer/providers/amazon/aws/triggers/sagemaker.py`
4. Regarding the unit tests, since we only extend the upstream provider and 
set `deferrable=True`, we only assert that the deferrable param is set in the 
Async operator by initialising the operator. All the other tests are removed
as the code no longer resides here and they would be covered in the upstream
provider. Hence, we're also removing the hook and  
trigger test modules.
5. Since the deferrable capability became available for these operators starting 
`apache-airflow-providers-amazon 8.1.0`, we set the minimum version for the 
provider in `setup.cfg` to `8.1.0`


closes: #757