Run EMR DAGs in parallel to reduce Master DAG run duration
Currently, we chain the EMR DAGs, the EMR EKS PI Job takes between
30 to 40 mins and the EMR Sensor DAGs takes about 8-10 mins. As a
result, due to these long running DAGs, the Master DAG takes
38 to 50 mins. The EMR EKS DAG takes the longest duration across
all the DAGs in Master DAG. With parallelising this chain, we're
expecting the Master DAG will run for a duration at most the same
as EMR EKS PI job and then it does not have to wait for addtional
8-10 mins for EMR sensor DAG to finish as it would be run parallelly
and would have completed by then.

The commit also removes additional free lines in the Slack report
as it looks unnecessary.