Delete stale IAM roles left by EMR EKS PI job DAG (#1191)
Ideally, the "emr_eks_pi_job" DAG should delete the IAM roles it
creates. However, there were observed occurrences where the IAM
roles were not deleted, most probably because we did not allow to
run the delete cluster task to completion or did not allow the
task to run altogether due to killing of the deployment before
it finished running. With this commit, we add a script that automates
the deletion of such stale IAM roles by first detaching its associated
policies. We call this script as a task from the AWS nuke DAG which is
scheduled to run after the end of the master DAG run.