Automate deletion of RC testing branch once PR is closed (#1196)
Our current GitHub configuration automatically deletes the branches
when PRs are merged. But for closed PRs, the branches are not deleted
and they remain stale. We have cases like testing the RCs, where we do
not want to merge the PR but only close it after verifying that our RCs
do not cause static check or tests failure. For such RC testing branches, 
this action will delete the branches upon closing the PR. 

In case, we want to reopen the PR, we need to click the button to 
restore the branch first and only then the reopen PR option will be
enabled by GitHub. I made an attempt to restore the branch if the event
is of type PR reopen, but that event cannot happen when the branch
no longer exists. Hence, we will have to manually click on the restore
button first if we want to reopen the PR.