Revert "feat(kubernetes): check state before deferring KubernetesPodOperatorAsync (#1104)"
This reverts commit 89ccc7ec1ad5c38a1665b50f649e9153d3ca2f37.

PR #1104 adds logic to poke for Pod status before putting the status
check on deferral. However, it is observed that our DAG fails always
when we go on checking the status.phase on pods as Pod.status is None
before the pod gets scheduled on a None. So, in most scenarios it
does not make sense to check for the pod status immediately to verify
that the pod has completed it's desired execution and hence, we revert
this poke in case of Kubernetes Pod operator.