Adding region option in eksctl commands (#1233)
It appears that the default region support in the latest release of eksctl (v0.148.0) is broken. A defect has already been raised for this issue, which you can find here. As a result of this problem, all of our eksctl commands were failing because we didn't have the "--region" command line option included. To work around this issue, I have temporarily added the "--region" option explicitly

Co-authored-by: rvatsastro <102275369+rvatsastro@users.noreply.github.com>