Adding task log verification in Master DAG (#1246)
In a recent release (6.2.0) of the airflow provider for Azure, we have observed that task logs are not visible in the Astro Cloud deployment with Azure Cluster.

To address this issue, we propose the following changes:

Add a task that checks the logs of the airflow version task. This will help us validate whether the logs are available in the Astro deployment.

Modify the Slack notification message to include cloud provider information.

Currently, our notification logic only displays the details of the example_dags that we have triggered. However, if any of the tasks in the master dag fail, we are not being notified. It is important to implement this now so that if the check_logs task fails, we can notify the failure in our slack report.

I have added airflow_dag_introspection.py which has some methods to get logs of a task, we can put this file in our deployment plugins folder to get this working