Execute Airflow Providers RC tests with a GCP cluster deployment (#1268)
As per the current process of providers' release testing, our workflow only updates the AWS deployment with RC Providers versions for testing. However, we want this update to be executed across all cloud providers. Currently, we are using a hosted organization, and Azure is not supported. We will enable this feature for Azure when it becomes available.