Remove sync wait for deletion of EKS cluster in EMR EKS DAG (#1307)
It is observed that the EKS cluster DAGs sometimes hangs on
the CloudFormation Stack to be deleted. It just prolongs the
task from it completion. Removing the sync wait as we could rely
on it finishing async and if it fails, the Nuke DAG can take care
of the deletion.