Pin max versions for CNCF Kubernetes and Google Airflow providers
The release of CNCF Kubernetes Airflow provider 7.9.0 includes
PR https://github.com/apache/airflow/pull/35422 whih removes
the PodLoggingStatus object that was getting returned that
we are relying in our Async KPO operator. Since, it is a breaking
change for our operator restrict the max version of the provider
to 7.8.0 until which the operator works well for us.
Additionally, since the Google KE engine operator also relies
on the PodLoggingStatus, also restrict the max version for the
Google provider to the previous stable version and restrict
the dependent Kubernetes provider version to the one containing
the PodLoggingStatus object