Revert "Pin max versions for CNCF Kubernetes and Google Airflow providers (#1359)"
This reverts commit 76d082c502a391bdb704e548cc56c1a421e2ffa3.
Since even if we pin the max version, there could be a possibility
that the cncf kubernetes provider version is upgraded in the deployment
and hence the pin won't be effective. It was decided to revert
the removal of required PodLoggingStatus object in Airflow OSS
and corresponding PR has been merged which can be tracked in the
linked issue.

closes: #1366