Correct google cloud-sdk download URL protocol to https from http
Recently, we have been observing failures to build our integration
tests docker images precisely at the step where we download the
google cloud-sdk library. It has been observed that the previously
working http URL is no longer available and we need to use a https
URL. This PR corrects the same for successful build of the image.

closes: #1488