Fix issue with API response not containing all tags for runtime version update (#1514)
We noticed that the action job to update the runtime version in the integration test deployment created a PR in which the version was not the latest. After debugging, we found out that the response from the API call does not contain all the tags. This PR addresses the issue after adding a limit query parameter